refactor: typo shema -> schema
